refactor(historia): refactor payment handling
