--- .changeset/2025-12-09-chore-convertoapi-docsite-web-markdown-s-sdk.md
@@ -1,5 +1,0 @@
----
-"@eventuras/event-sdk": patch
----
-
-- 🧹 chore(convertoapi,docsite,web,markdown,scribo,sdk): upgrade deps (b2de638) [@eventuras/event-sdk]
