+++ apps/historia/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eventuras/historia",
-  "version": "0.10.0",
+  "version": "0.11.0",
   "description": "Website template for Payload",
   "license": "MIT",
   "type": "module",
