+++ apps/web-e2e/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-e2e
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [7cbfcd4]
+  - @eventuras/event-sdk@0.6.1
+
 ## 0.1.1
 
 ### Patch Changes
