+++ apps/web-e2e/package.json
@@ -1,6 +1,6 @@
 {
   "name": "web-e2e",
-  "version": "0.1.1",
+  "version": "0.1.2",
   "private": "true",
   "description": "End to end tests for the Eventuras frontend",
   "keywords": [
