+++ libs/event-sdk/CHANGELOG.md
@@ -1,21 +1,30 @@
 # @eventuras/event-sdk
 
+## 0.6.1
+
+### Patch Changes
+
+- 7cbfcd4: - 🧹 chore(convertoapi,docsite,web,markdown,scribo,sdk): upgrade deps (b2de638) [@eventuras/event-sdk]
+
 ## 0.6.0
 
 ### Minor Changes
 
 ### 🧱 Features
-  - feat(web,event-sdk): await sdk-config (945510c) [@eventuras/event-sdk]
+
+- feat(web,event-sdk): await sdk-config (945510c) [@eventuras/event-sdk]
 
 ### 🧹 Maintenance
-  - chore(event-sdk): update with registration patch (0d3904e) [@eventuras/event-sdk]
-  - chore(event-sdk): rename client to client-next (9267bf7) [@eventuras/event-sdk]
-  - chore(event-sdk): minor comments formatting (8968091) [@eventuras/event-sdk]
+
+- chore(event-sdk): update with registration patch (0d3904e) [@eventuras/event-sdk]
+- chore(event-sdk): rename client to client-next (9267bf7) [@eventuras/event-sdk]
+- chore(event-sdk): minor comments formatting (8968091) [@eventuras/event-sdk]
 
 ### Patch Changes
 
 ### 🧹 Maintenance
-  - chore(convertoapi,docsite,web,markdown,scribo,sdk): upgrade deps (b2de638) [@eventuras/event-sdk]
+
+- chore(convertoapi,docsite,web,markdown,scribo,sdk): upgrade deps (b2de638) [@eventuras/event-sdk]
 
 ## 0.5.0
